ftpd now running and can be connected to through nat
Bit of a faff because there we can't use tun/tap inside the container
(as this would require privileged mode). The simh nat stack works
instead, but it also uses the same subnet as docker/podman
(10.0.2.0/24). This is ok though, and actually useful in some ways,
because it allows us to open an ACTIVE ftp session without trying to
set the source address.